web_app/pyproject.toml
[tool.poetry]
name = "rd-hw-9"
version = "0.1.0"
description = ""
authors = ["undefined <azmolwebautomation@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["all"], version = "^0.112.2"}
asyncpg = "^0.29.0"
sqlalchemy = "^2.0.32"
alembic = "^1.13.2"
gunicorn = "^23.0.0"
uvicorn = "^0.30.6"
aiofiles = "^24.1.0"
pytest-asyncio = "^0.24.0"
pydantic = "^2.8.2"
pytest-order = "^1.3.0"
psycopg2 = "^2.9.9"
celery = {extras = ["redis"], version = "^5.4.0"}
gitpython = "^3.1.43"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
